split System and Cell, update __init__, rm unused
