+++ src/topos/__init__.py
@@ -1,5 +1,3 @@
 from .set import Set, Iterable, Map, Mapping
 from .hypergraph import Hypergraph, Chain
-from .product import Prod
-from .tensor import Tensor, Matrix, Oplus
-from .timed import timed
+from .shape import Shape, Cell
